# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * soukMA - Moroccan e-commerce marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessEnvelope {
  success: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export type MyProfileRole = (typeof MyProfileRole)[keyof typeof MyProfileRole];

export const MyProfileRole = {
  guest: "guest",
  customer: "customer",
  vendor: "vendor",
  admin: "admin",
} as const;

export interface Vendor {
  id: string;
  shopName: string;
  /** @nullable */
  description?: string | null;
  phone: string;
  city: string;
  createdAt: string;
}

export interface MyProfile {
  authenticated: boolean;
  role: MyProfileRole;
  user?: AuthUser | null;
  vendor?: Vendor | null;
}

export interface BecomeVendorBody {
  /**
   * @minLength 2
   * @maxLength 80
   */
  shopName: string;
  /** @maxLength 500 */
  description?: string;
  /**
   * @minLength 6
   * @maxLength 30
   */
  phone: string;
  /**
   * @minLength 2
   * @maxLength 80
   */
  city: string;
}

export interface Category {
  id: string;
  slug: string;
  name: string;
  /** @nullable */
  nameAr?: string | null;
  /** @nullable */
  icon?: string | null;
  productCount: number;
}

export interface Product {
  id: string;
  slug: string;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  currency: string;
  stock: number;
  categorySlug: string;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  vendorId?: string | null;
  vendorName: string;
  /** @nullable */
  imageUrl: string | null;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  createdAt: string;
}

export interface ProductList {
  items: Product[];
  total: number;
}

export interface ProductSpec {
  label: string;
  value: string;
}

export type ProductDetail = Product & {
  images: string[];
  specs: ProductSpec[];
};

export interface CreateProductBody {
  /**
   * @minLength 2
   * @maxLength 160
   */
  title: string;
  /** @maxLength 4000 */
  description?: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  compareAtPrice?: number;
  /** @minimum 0 */
  stock: number;
  categorySlug: string;
  imageUrl?: string;
}

export interface UpdateProductBody {
  /**
   * @minLength 2
   * @maxLength 160
   */
  title?: string;
  /** @maxLength 4000 */
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  compareAtPrice?: number;
  /** @minimum 0 */
  stock?: number;
  categorySlug?: string;
  imageUrl?: string;
  isFeatured?: boolean;
}

export interface CartItem {
  id: string;
  productId: string;
  productTitle: string;
  /** @nullable */
  productImage: string | null;
  unitPrice: number;
  quantity: number;
  lineTotal: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  currency: string;
  itemCount: number;
}

export interface AddCartItemBody {
  productId: string;
  /**
   * @minimum 1
   * @maximum 99
   */
  quantity: number;
}

export interface UpdateCartItemBody {
  /**
   * @minimum 1
   * @maximum 99
   */
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  paid: "paid",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cmi: "cmi",
  cih_pay: "cih_pay",
  cash_on_delivery: "cash_on_delivery",
} as const;

export interface Order {
  id: string;
  reference: string;
  status: OrderStatus;
  total: number;
  currency: string;
  itemCount: number;
  paymentMethod: OrderPaymentMethod;
  createdAt: string;
}

export interface OrderItem {
  id: string;
  productId: string;
  productTitle: string;
  /** @nullable */
  productImage: string | null;
  unitPrice: number;
  quantity: number;
  lineTotal: number;
  vendorName: string;
}

export interface ShippingAddress {
  fullName: string;
  phone: string;
  address: string;
  city: string;
  /** @nullable */
  notes?: string | null;
}

export type OrderDetail = Order & {
  items: OrderItem[];
  shipping: ShippingAddress;
};

export type PlaceOrderBodyPaymentMethod =
  (typeof PlaceOrderBodyPaymentMethod)[keyof typeof PlaceOrderBodyPaymentMethod];

export const PlaceOrderBodyPaymentMethod = {
  cmi: "cmi",
  cih_pay: "cih_pay",
  cash_on_delivery: "cash_on_delivery",
} as const;

export interface PlaceOrderBody {
  shipping: ShippingAddress;
  paymentMethod: PlaceOrderBodyPaymentMethod;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  paid: "paid",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface VendorStats {
  totalProducts: number;
  totalOrders: number;
  revenue: number;
  currency: string;
  lowStockCount: number;
  recentOrders: Order[];
  topProducts: Product[];
}

export type AdminStatsOrdersByStatusItem = {
  status: string;
  count: number;
};

export type AdminStatsTopCategoriesItem = {
  slug: string;
  name: string;
  productCount: number;
};

export interface AdminStats {
  totalUsers: number;
  totalVendors: number;
  totalProducts: number;
  totalOrders: number;
  revenue: number;
  currency: string;
  ordersByStatus: AdminStatsOrdersByStatusItem[];
  recentOrders: Order[];
  topCategories: AdminStatsTopCategoriesItem[];
}

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  customer: "customer",
  vendor: "vendor",
  admin: "admin",
} as const;

export interface AdminUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  role: AdminUserRole;
  createdAt: string;
}

export interface InitiatePaymentBody {
  orderId: string;
}

export type PaymentSessionProvider =
  (typeof PaymentSessionProvider)[keyof typeof PaymentSessionProvider];

export const PaymentSessionProvider = {
  cmi: "cmi",
  cih_pay: "cih_pay",
} as const;

export interface PaymentSession {
  orderId: string;
  provider: PaymentSessionProvider;
  redirectUrl: string;
  sessionId: string;
}

export interface ConfirmPaymentBody {
  orderId: string;
  sessionId: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListProductsParams = {
  search?: string;
  categorySlug?: string;
  minPrice?: number;
  maxPrice?: number;
  sort?: ListProductsSort;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minimum 0
   */
  offset?: number;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  newest: "newest",
  priceAsc: "priceAsc",
  priceDesc: "priceDesc",
  popular: "popular",
} as const;