# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: soukMA - Moroccan e-commerce marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      tags: [Products]
      operationId: listProducts
      summary: List products (search, filter, sort)
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: categorySlug
          in: query
          required: false
          schema: { type: string }
        - name: minPrice
          in: query
          required: false
          schema: { type: number }
        - name: maxPrice
          in: query
          required: false
          schema: { type: number }
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [newest, priceAsc, priceDesc, popular]
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
...
  /products/{id}:
    get:
      tags: [Products]
      operationId: getProduct
      summary: Get product by id with full details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /payment/initiate:
    post:
      tags: [Payment]
      operationId: initiatePayment
      summary: Initiate an online payment (CMI / CIH Pay placeholder)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiatePaymentBody'
      responses:
        '200':
          description: Payment session info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSession'
  /payment/confirm:
    post:
      tags: [Payment]
      operationId: confirmPayment
      summary: Confirm payment (mock for CMI / CIH Pay)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmPaymentBody'
      responses:
        '200':
          description: Confirmed order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Product:
      type: object
      required: [id, slug, title, price, currency, stock, categorySlug, vendorName, imageUrl, rating, reviewCount, isFeatured, createdAt]
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        description:
          type: [string, 'null']
        price: { type: number }
        compareAtPrice:
          type: [number, 'null']
        currency: { type: string }
        stock: { type: integer }
        categorySlug: { type: string }
        categoryName:
          type: [string, 'null']
        vendorId:
          type: [string, 'null']
        vendorName: { type: string }
        imageUrl:
          type: [string, 'null']
        rating: { type: number }
        reviewCount: { type: integer }
        isFeatured: { type: boolean }
        createdAt: { type: string, format: date-time }
    ProductList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total: { type: integer }
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          required: [images, specs]
          properties:
            images:
              type: array
              items: { type: string }
            specs:
              type: array
              items:
                $ref: '#/components/schemas/ProductSpec'
    ProductSpec:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
...
    Order:
      type: object
      required: [id, reference, status, total, currency, itemCount, paymentMethod, createdAt]
      properties:
        id: { type: string }
        reference: { type: string }
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        total: { type: number }
        currency: { type: string }
        itemCount: { type: integer }
        paymentMethod:
          type: string
          enum: [cmi, cih_pay, cash_on_delivery]
        createdAt: { type: string, format: date-time }
    OrderItem:
      type: object
      required: [id, productId, productTitle, productImage, unitPrice, quantity, lineTotal, vendorName]
      properties:
        id: { type: string }
        productId: { type: string }
        productTitle: { type: string }
        productImage:
          type: [string, 'null']
        unitPrice: { type: number }
        quantity: { type: integer }
        lineTotal: { type: number }
        vendorName: { type: string }
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          required: [items, shipping]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            shipping:
              $ref: '#/components/schemas/ShippingAddress'
    ShippingAddress:
      type: object
      required: [fullName, phone, address, city]
      properties:
        fullName: { type: string }
        phone: { type: string }
        address: { type: string }
        city: { type: string }
        notes:
          type: [string, 'null']
...
    InitiatePaymentBody:
      type: object
      required: [orderId]
      properties:
        orderId: { type: string }
    PaymentSession:
      type: object
      required: [orderId, provider, redirectUrl, sessionId]
      properties:
        orderId: { type: string }
        provider:
          type: string
          enum: [cmi, cih_pay]
        redirectUrl: { type: string }
        sessionId: { type: string }
    ConfirmPaymentBody:
      type: object
      required: [orderId, sessionId]
      properties:
        orderId: { type: string }
        sessionId: { type: string }