# lib/api-zod/src/generated/types/paymentSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * soukMA - Moroccan e-commerce marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentSessionProvider } from "./paymentSessionProvider";

export interface PaymentSession {
  orderId: string;
  provider: PaymentSessionProvider;
  redirectUrl: string;
  sessionId: string;
}