# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * soukMA - Moroccan e-commerce marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  slug: string;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  currency: string;
  stock: number;
  categorySlug: string;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  vendorId?: string | null;
  vendorName: string;
  /** @nullable */
  imageUrl: string | null;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  createdAt: Date;
}